fastapi-backend/requirements.txt
bcrypt 
cryptography
fastapi
python-multipart
python-jose
sqlalchemy 
uvicorn
passlib